Fixed NPE when using EventTarget.detach('cat|*'), with a prefix configured.
The exception was coming from the private `_getType(type, pre)` method.

We removed the typeof string check, for performance reasons in 3.10.0, since the
calling code had it covered already. However there's one path, for the 'cat|*' star case,
where type === null (for '*'), and it was barfing on the t.indexOf() call as a result.