Fixed NPE on EventTarget.detach('cat|*')
The exception was coming from the private `_getType(type, pre)` method.

We removed the typeof string check in `getType`, for performance reasons
in 3.10.0, since the calling code had it covered already and it sits on many
critical paths.

However there's one path, for the 'cat|*' star case with a prefix configured,
where a null `type` is passed to `_getType` and it was barfing on the `type.indexOf()`
call as a result.

Re-added a simple truthy check (still avoiding the typeof string) before calling
indexOf.