resolves #1062 - fix loader test that checks for bleeding dependencies
We only check that the arrays contain the same items. We do not care
about the order of modules in the flattened dependency tree since that's
not guaranteed to be the same across browsers.

We now include nodejs and IE8 since that was the original UA coverage
intent of this test.

Comments++!