Update proxy.js
There is no way to modify the cursor style on the drag node, unless the initial cursor style is set to "auto". I would like to be able to customize both the initial hovering cursor style as well as the drag cursor style.

I find the whole dragCursor implementation to me clunky to begin with but at least this allows me to use the attribute when I actually want to.