date: Properly ignore tests with bad dynamic input
Certain tests inside of the Date tests need to be ignored in cases
where `(new Date()).toString()` is faulty.

This is a conscious choice to allow the tests to successfully run in
all timezones, and do not affect the correctness of the code.

It's a trade-off that needs to be made - whether the tests can run in
all browsers, or whether the tests can run in all timezones.