Ignore running some tests by Yeti in Calendar tests
Currently, running Calendar tests containing the focus process will
fail in Firefox. The cause of this problem seems to be a problem that
test case is not bad, what happens if you are running Yeti while running
Firefox in the background. This problem does not occur in a browser
other than Firefox. In addition, there is a test that includes a focusing
process in the same way the Widget Button module [1], but it has to
ignore Yeti is here. Therefore, I would like to ignore running some tests
by Yeti in the same way in Calendar. Fix #1363.

[1]: src/widget-buttons/tests/unit/assets/widget-buttons-test.js