Convert Y.log to Y.message for deprecations
This has the result that deprecation warnings are not stripped by Shifter
by default such that it is possible to run the raw or minified versions of
the core YUI library when developing non-core code and still see
deprecation warnings.

For production usage, Y.config.debug should be set to false to ensure that
all messages are removed from output.