[dom] Normalize setting a node value to null
When set a node value to `null`, then a node value returns `"null"` as string
on the legacy IE (IE <= 9). A node value returns an empty string on the modern
browsers (Chrome, Firefox, IE >= 10). This change is to normalize the results
of setting a node value to `null` on any browsers. Fix #1603.