[core] Fix coverage tests
The original coverage tests used yui-core-coverage and loaded the rest
of yui-base by inserting script tags. This seems to have been set up
that way to get more granular coverage information than just one entry
for yui-base, but it is causing YUI.add to fail in certain cases because
Loader is not yet available. This changes the coverage test to do the
same as the regular unit tests.