Move preventDefault to gesturemovestart (Fixes #1716)
Android browsers require that we prevent the default browser action on the
start of drag. We tried this in #693/#1557, but this broke in other ways
for other browsers.

The issue stems from preventDefault being called on the document instead of
the Node.

This commit moves the preventDefault to the gesturemovestart in the
prototype for DD.Drag.