Fix #1816: ScrollViewPaginator partial transition
This is fixing the issue described in #1816, which causes diagonal flicks to change the paginator's index even though the image showing up is not updated.
To fix this I'm just adding one more check to _beforeHostFlick to ignore flicks that don't happen in the specified flick axis. This is the default behavior of scrollview-base, but paginator-plugin prevents it and runs its own code, so we need the check there as well.