Fix #1740: Scrollview scrolltoindex breaks after a click-drag down
I've come across the problem described in #1740 and found out that this is happening because _beforeHostScrollTo is always trying to adjust the scroll coordinates according to the last gesture that was made.
It's possible that the scrolling wasn't caused by a gesture though, like on direct calls to scrollToIndex (the buttons on http://yuilibrary.com/yui/docs/scrollview/scrollview-paging-example.html for example). And if this is done right after a gesture which is not on the scrolling axis, _beforeHostScrollTo will clear the coordinates, which is why the index is being changed but the new image is not being scrolled into view.
I've fixed this by making sure that _beforeHostScrollTo doesn't try using the last gesture to adjust the coordinates for a call to scrollToIndex, since this call will already have the right coordinate anyway.