Fix #1901: stopPropagation doesn't work for gesturemovestart when minTime or minDistance is set
When using minTime or minDistance, stopPropagation calls don't work anymore for the gesturemovestart event. That's because these are triggered asynchronously, so stopPropagation ends up being called after the original events already propagated. This is explained in more detail in issue #1901.
This is fixing the problem by adding a flag to the original event indicating it has been stopped, and then preventing any gesturemovestart events tied events with that flag from firing.